requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-classic
groq
transformers
torch
chromadb
requests
beautifulsoup4
playwright
pymupdf
docx2txt
python-multipart
python-multipart
pydantic[email]
aiosmtplib
python-dotenv